Extract `FeedViewController` test helpers to a separate file
